requirements.txt
pillow==5.4.1
qrcode==6.1
gunicorn==20.0.4
requests==2.23.0
python-dateutil==2.6.0
geographiclib==1.50
psycopg2-binary==2.8.4
pypdf2==1.26
django==2.2.5
django-taggit==1.2.0
django-tinymce==2.8.0
django-nested-inline==0.4.2
django-autocomplete-light==3.5.1
django-auditlog==0.4.7
django-background-tasks==1.2.5
django-debug-toolbar==2.2
django-sendgrid-v5==0.8.1
django-multiselectfield==0.1.12
django-session-timeout==0.1.0
fcm-django==0.3.4
